+++ packages/nx-plugin/src/generators/init/schema.d.ts
@@ -1,0 +1,8 @@
+export interface SetupAnalogGeneratorSchema {
+  project: string;
+  vitest: boolean;
+}
+
+export interface NormalizedSchema extends SetupAnalogGeneratorSchema {
+  projectRoot: string;
+}