+++ packages/nx-plugin/src/utils/versions/ng_16_X/versions.ts
@@ -1,0 +1,12 @@
+// V16_X
+export const V16_X_ANALOG_JS_ROUTER = '^1.2.0';
+export const V16_X_ANALOG_JS_CONTENT = '^1.2.0';
+
+// devDependencies
+export const V16_X_ANALOG_JS_PLATFORM = '^1.2.0';
+export const V16_X_ANALOG_JS_VITE_PLUGIN_ANGULAR = '^1.2.0';
+export const V16_X_NX_ANGULAR = '~18.0.0';
+export const V16_X_NX_VITE = '~18.0.0';
+export const V16_X_JSDOM = '^22.0.0';
+export const V16_X_VITE_TSCONFIG_PATHS = '^4.2.0';
+export const V16_X_VITEST = '^1.31.0';